pyproject.toml
[project]
name = "snipster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Karthik Ghantasala", email = "karthikghan@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
]

[project.scripts]
snipster = "snipster:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "poethepoet>=0.36.0",
    "pytest-cov>=6.2.1",
]

[tool.poe.tasks]
test = "pytest"
cov = "pytest --cov=commonactors --cov-report=term-missing"
lint = "ruff check ."
format = "ruff format ."
